Refactor: Min/Max Calculations
Before:
Had a function whose job was to generate "all the min/max data". It
created a "minMaxFinder" for each time unit, stuck them in an array,
looped through subtotals and the finders to generate data, then returned
the results in an object.

Now:
Individual function for each time unit, which calls minMaxFinder with
the appropriate config. The loop through subtotals is now within the
minMaxFinder.

Results:
* Decoupled config from calculations. If someone wants to remove a time
  unit or add a new one, it's now much easier
* Now possible to calculate min/max for just one time unit
* Simplified config / hid implementation details by running all updates
  through the same updater function, which determines if "existing" or
  "working" is appropriate
